OSIS-3949 moved AuthorizedRelationshipList object from program_management/ddd/domain into base/models/authorized_relationship
