+++ program_management/ddd/domain/program_tree.py
@@ -27,7 +27,7 @@ from typing import List, Set
 
 from base.models.enums.education_group_types import EducationGroupTypesEnum, TrainingType
 from program_management.ddd.business_types import *
-from program_management.ddd.domain.authorized_relationship import AuthorizedRelationshipList
+from base.models.authorized_relationship import AuthorizedRelationshipList
 from program_management.ddd.validators.validators_by_business_action import AttachNodeValidatorList
 
 PATH_SEPARATOR = '|'
