+++ program_management/tests/ddd/domain/test_authorized_relationship.py
@@ -26,7 +26,7 @@
 from django.test import SimpleTestCase
 
 from base.models.enums.education_group_types import TrainingType, GroupType
-from program_management.ddd.domain.authorized_relationship import AuthorizedRelationshipList
+from base.models.authorized_relationship import AuthorizedRelationshipList
 from program_management.tests.ddd.factories.authorized_relationship import AuthorizedRelationshipFactory
 from program_management.tests.ddd.factories.node import NodeGroupYearFactory
